# Optimization improvements