So from discussion with @NikVolf it seems like some of our benchmarks were rather unrepresentative, so I've fixed up the multiplication code to less aggressively prefer 0 and other small numbers. We're now at faster-than-asm performance for the full multiplication and only slightly slower for overflowing multiplication when done over the full range of numbers (product of squares of all odd numbers between 0 and 10000, or more specifically the product of `(0..1000).map(|x| x | 1)`). I'm wondering whether we can make the overflowing multiplication slightly faster by adding a check in the loop something similar to

```rust
if index_about_to_be_set > $out_words / 2 && num_about_to_be_set != 0 {
    is_overflowing = true;
}
```

Maybe we'll want to put `&& !is_overflowing` in the condition to get better branch prediction, but we'd have to instrument it to make sure. The `index_about_to_be_set` check should be optimized out by const-folding.

Anyway, this should be an improvement over the multiplication that was included in the previous PR.

## Benchmarks:

[Old PR](https://github.com/paritytech/bigint/pull/26):

```
running 10 tests
test u128_mul      ... bench:      85,756 ns/iter (+/- 1,131)
test u256_add      ... bench:      33,537 ns/iter (+/- 17,619)
test u256_from_be  ... bench:          25 ns/iter (+/- 0)
test u256_from_le  ... bench:          11 ns/iter (+/- 1)
test u256_full_mul ... bench:     307,933 ns/iter (+/- 2,363)
test u256_mul      ... bench:     334,596 ns/iter (+/- 11,623)
test u256_sub      ... bench:      37,168 ns/iter (+/- 3,088)
test u512_add      ... bench:      30,430 ns/iter (+/- 509)
test u512_mul      ... bench:   1,387,942 ns/iter (+/- 158,007)
test u512_sub      ... bench:      31,861 ns/iter (+/- 571)
```

asm:

```
running 10 tests
test u128_mul      ... bench:     287,110 ns/iter (+/- 2,359)
test u256_add      ... bench:      43,413 ns/iter (+/- 668)
test u256_from_be  ... bench:          25 ns/iter (+/- 0)
test u256_from_le  ... bench:          11 ns/iter (+/- 0)
test u256_full_mul ... bench:     231,655 ns/iter (+/- 1,247)
test u256_mul      ... bench:     124,843 ns/iter (+/- 33,822)
test u256_sub      ... bench:      44,612 ns/iter (+/- 225)
test u512_add      ... bench:     107,818 ns/iter (+/- 656)
test u512_mul      ... bench:     806,559 ns/iter (+/- 64,862)
test u512_sub      ... bench:     174,912 ns/iter (+/- 21,225)
```

This PR:

```
running 10 tests
test u128_mul      ... bench:      76,520 ns/iter (+/- 6,218)
test u256_add      ... bench:      33,468 ns/iter (+/- 2,700)
test u256_from_be  ... bench:          25 ns/iter (+/- 0)
test u256_from_le  ... bench:          11 ns/iter (+/- 1)
test u256_full_mul ... bench:     212,304 ns/iter (+/- 18,580)
test u256_mul      ... bench:     144,798 ns/iter (+/- 3,291)
test u256_sub      ... bench:      37,195 ns/iter (+/- 722)
test u512_add      ... bench:      30,443 ns/iter (+/- 2,172)
test u512_mul      ... bench:     211,607 ns/iter (+/- 4,644)
test u512_sub      ... bench:      31,866 ns/iter (+/- 582)
```

Benchcmp (old PR vs this PR):

```
 name           old.bench ns/iter  optimized.bench ns/iter  diff ns/iter   diff %  speedup 
 u128_mul       85,756             76,520                         -9,236  -10.77%   x 1.12 
 u256_add       33,537             33,468                            -69   -0.21%   x 1.00 
 u256_from_be   25                 25                                  0    0.00%   x 1.00 
 u256_from_le   11                 11                                  0    0.00%   x 1.00 
 u256_full_mul  307,933            212,304                       -95,629  -31.06%   x 1.45 
 u256_mul       334,596            144,798                      -189,798  -56.72%   x 2.31 
 u256_sub       37,168             37,195                             27    0.07%   x 1.00 
 u512_add       30,430             30,443                             13    0.04%   x 1.00 
 u512_mul       1,387,942          211,607                    -1,176,335  -84.75%   x 6.56 
 u512_sub       31,861             31,866                              5    0.02%   x 1.00 
```

Benchcmp (asm vs this PR):

```
 name           asm.bench ns/iter  optimized.bench ns/iter  diff ns/iter   diff %  speedup 
 u128_mul       287,110            76,520                       -210,590  -73.35%   x 3.75 
 u256_add       43,413             33,468                         -9,945  -22.91%   x 1.30 
 u256_from_be   25                 25                                  0    0.00%   x 1.00 
 u256_from_le   11                 11                                  0    0.00%   x 1.00 
 u256_full_mul  231,655            212,304                       -19,351   -8.35%   x 1.09 
 u256_mul       124,843            144,798                        19,955   15.98%   x 0.86 
 u256_sub       44,612             37,195                         -7,417  -16.63%   x 1.20 
 u512_add       107,818            30,443                        -77,375  -71.76%   x 3.54 
 u512_mul       806,559            211,607                      -594,952  -73.76%   x 3.81 
 u512_sub       174,912            31,866                       -143,046  -81.78%   x 5.49 
```