@Vurich 2017-08-21
@NikVolf Yes. Don't merge this yet, the benchmarks were ran on buggy code and I'm still working on it.