@Vurich 2017-08-21
So the optimisation work I was doing was good, but overall slower than the naive method that we already had. Unrolling the loops helped but the fancier stuff I was trying slowed down larger numbers too much for me to be comfortable merging this. I think getting the Rust implementation of multiplication down to par with the asm is blocked on stablising 128-bit arithmetic, since our software implementation of 128-bit arithmetic is extremely inefficient. Another alternative would be to switch to always using `u32` instead of `u64`, but although that would speed up multiplication it would require a penalty on every other operation. I'll close this PR and reopen another one with only the improved benchmarks.